requirements.txt
python-telegram-bot>=21.0,<22.0
python-dotenv>=1.0.0
openai>=1.40.0
rapidfuzz>=3.0.0,<4.0.0
